chore(github): Update QT install action to use internal cache
v2 of the `install-qt-action` action requires v1 of the `cache` action
which does not support caching manual trigger `workflow_dispatch`
runs (and IIRC schedule triggers also).

Thus meaning the cache was being lost for the 32bit builds since it
would never get a push or PR trigger to keep it from expiring.

v3 supports internal caching, so this commit aims to enable and test it.